% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transcripts.R
\name{getSeqLengths}
\alias{getSeqLengths}
\title{Returns sequence lengths from sequences in a FASTA file}
\usage{
getSeqLengths(fastaFile)
}
\arguments{
\item{fastaFile}{filepath of target FASTA file}
}
\description{
Returns sequence lengths from sequences in a FASTA file
}
\author{
Matthew Bernstein \email{matthewb@cs.wisc.edu}
}

