% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transcripts.R
\name{histSeqLengths}
\alias{histSeqLengths}
\title{Creates a histogram of the sequence lengths from sequences in a FASTA file}
\usage{
histSeqLengths(fastaFile)
}
\arguments{
\item{fastaFile}{filepath of target FASTA file}
}
\description{
Creates a histogram of the sequence lengths from sequences in a FASTA file
}
\author{
Matthew Bernstein \email{matthewb@c.wisc.edu}
}

